# src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: TopAgents API
  description: |
    ## TopAgents.ca — Real Estate Agent Review & Referral Platform

    ### Authentication
    Protected endpoints require a `Bearer` JWT token in the `Authorization` header.
    Get tokens via `POST /api/v1/auth/login`. Refresh via `POST /api/v1/auth/refresh`.

    ### Rate Limiting
    - General: 100 req / 15 min
    - Auth endpoints: 10 req / 15 min
    - Search: 60 req / 1 min

    ### Pagination
    All list endpoints accept `?page=1&limit=20`.
    Response includes `meta: { total, page, limit, totalPages, hasNext, hasPrev }`.

    ### Response Envelope
    ```json
    { "success": true, "data": { ... } }
    ```
    ```json
    { "success": false, "error": { "code": 422, "message": "Validation failed", "details": [] } }
    ```
  version: 1.0.0
  contact:
    name: TopAgents Dev Team
    email: dev@topagents.ca
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.topagents.ca/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        license_number:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        brokerage:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        languages:
          type: array
          items:
            type: string
        specialties:
          type: array
          items:
            type: string
        areas_served:
          type: array
          items:
            type: string
        claimed_at:
          type: string
          format: date-time
          nullable: true
        verified_at:
          type: string
          format: date-time
          nullable: true
        current_tier:
          type: string
          enum: [BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, LUMINARY]
        total_points:
          type: integer
        year_points:
          type: integer
        avg_rating:
          type: number
          format: float
        total_reviews:
          type: integer
        connected_platforms:
          type: integer
        completed_referrals:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentCreate:
      type: object
      required: [license_number, first_name, last_name]
      properties:
        license_number:
          type: string
          maxLength: 50
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
        brokerage:
          type: string
        bio:
          type: string
          maxLength: 2000
        website:
          type: string
          format: uri
        languages:
          type: array
          items:
            type: string
        specialties:
          type: array
          items:
            type: string
        areas_served:
          type: array
          items:
            type: string
    AgentUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        brokerage:
          type: string
        bio:
          type: string
          maxLength: 2000
        photo_url:
          type: string
        website:
          type: string
          format: uri
        languages:
          type: array
          items:
            type: string
        specialties:
          type: array
          items:
            type: string
        areas_served:
          type: array
          items:
            type: string
        lead_notifications:
          type: boolean
        referral_notifications:
          type: boolean
        marketing_emails:
          type: boolean
    Review:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: integer
        platform:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          nullable: true
        text:
          type: string
          nullable: true
        reviewer_name:
          type: string
        verified:
          type: boolean
        transaction_verified:
          type: boolean
        review_date:
          type: string
          format: date
        agent_responded:
          type: boolean
        agent_response:
          type: string
          nullable: true
        response_date:
          type: string
          format: date-time
          nullable: true
        property_address:
          type: string
          nullable: true
        transaction_type:
          type: string
          nullable: true
        rating_communication:
          type: integer
          nullable: true
        rating_negotiation:
          type: integer
          nullable: true
        rating_market_knowledge:
          type: integer
          nullable: true
        rating_responsiveness:
          type: integer
          nullable: true
        rating_pricing:
          type: integer
          nullable: true
        rating_paperwork:
          type: integer
          nullable: true
        rating_availability:
          type: integer
          nullable: true
        rating_recommendation:
          type: integer
          nullable: true
        imported_at:
          type: string
          format: date-time
...
    City:
      type: object
      properties:
        id:
          type: integer
        city_name:
          type: string
        region:
          type: string
        province:
          type: string
        slug:
          type: string
        population:
          type: integer
          nullable: true
        avg_home_price:
          type: integer
          nullable: true
        total_agents:
          type: integer
        seo_title:
          type: string
          nullable: true
        seo_description:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        top_agents:
          type: array
          items:
            type: object
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /agents:
    get:
      tags: [Agents]
      summary: Search and filter agents
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search by name or brokerage
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: province
          schema:
            type: string
        - in: query
          name: min_rating
          schema:
            type: number
            minimum: 1
            maximum: 5
        - in: query
          name: tier
          schema:
            type: string
            enum: [BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, LUMINARY]
        - in: query
          name: specialties
          schema:
            type: string
          description: Comma-separated list
        - in: query
          name: languages
          schema:
            type: string
          description: Comma-separated list
        - in: query
          name: sort
          schema:
            type: string
            enum: [rating, reviews, recent, points]
            default: rating
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated list of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Agents]
      summary: Create agent profile (admin or bulk seed)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Agent created
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: License number already exists
  /agents/{agentId}:
    get:
      tags: [Agents]
      summary: Get full agent profile by ID
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Agent profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Agents]
      summary: Update agent profile (owner or admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Updated agent
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{agentId}/claim:
    post:
      tags: [Agents]
      summary: Claim an unclaimed agent profile
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [license_number]
              properties:
                license_number:
                  type: string
                  example: REA-123456
      responses:
        '200':
          description: Profile claimed successfully
        '400':
          description: License number does not match
        '409':
          description: Profile already claimed
  /agents/{agentId}/stats:
    get:
      tags: [Agents]
      summary: Get agent performance statistics
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Review stats, referral counts, tier and points summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      reviews:
                        type: object
                        properties:
                          total_reviews:
                            type: integer
                          avg_rating:
                            type: number
                          five_star:
                            type: integer
                          four_star:
                            type: integer
                          three_star:
                            type: integer
                          two_star:
                            type: integer
                          one_star:
                            type: integer
                          avg_communication:
                            type: number
                          avg_negotiation:
                            type: number
                          avg_market_knowledge:
                            type: number
                          avg_responsiveness:
                            type: number
                      referrals:
                        type: object
                        properties:
                          completed:
                            type: integer
                          sent_completed:
                            type: integer
                          received_completed:
                            type: integer
                      points:
                        type: object
                        properties:
                          total_points:
                            type: integer
                          year_points:
                            type: integer
                          current_tier:
                            type: string
  /agents/{agentId}/reviews:
    get:
      tags: [Agents]
      summary: Get paginated reviews for an agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
        - in: query
          name: rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - in: query
          name: platform
          schema:
            type: string
        - in: query
          name: verified
          schema:
            type: boolean
        - in: query
          name: sort
          schema:
            type: string
            enum: [recent, highest, lowest]
            default: recent
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
  /cities:
    get:
      tags: [Cities]
      summary: List all BC cities with agent counts
      parameters:
        - in: query
          name: province
          schema:
            type: string
            default: BC
        - in: query
          name: region
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paginated city list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/City'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /cities/{slug}:
    get:
      tags: [Cities]
      summary: Get city landing page data with top 6 agents
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
          example: burnaby
      responses:
        '200':
          description: City data including SEO fields and top agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/City'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── ADMIN ─────────────────────────────────────────────────────────────────
...
  /admin/agents:
    get:
      tags: [Admin]
      summary: List all agents including unclaimed profiles
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [claimed, unclaimed, all]
            default: all
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated full agent list
...